CallbackUrl and permission Scopes can now be saved with the env
The CallbackUrl and user permission Scopes can now be saved with the environment configuration. This allows a custom default CallbackUrl and Scopes to be used during Initialize-SpotifySession which for simplicity does not have all the parameters that its underlying function has (Initialize-SpotifyAuthorizationCodeFlow). This will reduce the number of things a user will have to set on first time setup if they want default settings.