Removing possibly unneeded copy of scopes during refresh
This code shouldn't really be needed as the scopes attribute should be returned on refreshes and full authentication/authorization process. The only case I have seen where the scopes attribute is not present is during a reauthentication process where you have already been approved and you are not specifying the need to re-authorize the scopes. This is described in the following issue #4 .