Reverting output type conversions
These conversions didn't work the way I expected. PowerShell pipeline will always take my typed arrays (i.e. string[]) and enumerate it and later recreate the array as an object[] array.

The only way around this is to put the typed array into another array with an "empty element" as the first array element (i.e. @(,$arr)). When the pipeline enumerates that array it skips the first "empty element" and outputs the typed array as is. Note: @(,$arr) is not the same as @($null, $arr).

However, this behavior would make function parameters that get their ValueByPipeline .....less simple to work with.